Update IO.hsc (#5)
* Update IO.hsc

On line 168, the third parameter for `netcode_server_receive_packet` is incorrectly marked as `Ptr Word8`. This parameter stores the size of the packet that is received by the server, and is marked as an `int` in the C code. Because it is marked as `Word8` in the Haskell binding, it can't store the size of packets bigger than 255 bytes, which results in truncated packets in the Haskell code.

This change fixes it.

* Update IO.hsc

Changed `Int` to `CInt`.